move updateMoodHistoryListener into DBMoodSetter
